requirements.txt
waitress
flask
python-dotenv
orator
orator_cache
psycopg2
redis